let ban duplicate classes ignore when the bytecode matches exactly
If we have two classes with the exact same name on the classpath BanDuplicateClasses fails, but with this enhancement it only fails if the bytecode of those class files differs. In other words, if they're an exact binary match we skip the classes and move on. The idea being: there's technically no problem if they're the same exact file. This also means there's potentially less fixes to make when porting a large pom.xml to start using BanDuplicateClasses for the first time. With this new option enabled any time BanDuplicateClasses fails it means there is a definite classpath problem, 100% of the time.

The new option is off by default, but can be enabled in pom files.